Have done with validation in Categories

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Http\Requests\CategoryRequest;
 use Illuminate\Support\File;
 use Illuminate\Support\Facades\Storage;
 
@@ -38,11 +39,16 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if($request->has('image'))
+        {
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
 
         Category::create($params);
 
@@ -78,13 +84,19 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if($request->has('image'))
+        {
+            Storage::delete($category->image);
+
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
+
 
         $category->update($params);
